refactor: rename main controllers
