+++ src/extension.ts
@@ -1,13 +1,13 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
-import { SystemProvider } from './systemProvider';
-import { UserProvider } from './userProvider';
+import { BuiltinController } from './builtinController';
+import { FileController } from './fileController';
 
 // this method is called when your extension is activated
 export function activate(context: vscode.ExtensionContext): void {
-    new SystemProvider(context);
-    new UserProvider(context);
+    new BuiltinController(context);
+    new FileController(context);
 }
 
 // this method is called when your extension is deactivated
